@@ -7,13 +7,18 @@
     "express": "^4.14.0",
     "mime": "^1.3.4",
     "pug": "^2.0.0-beta6",
-    "socket.io": "^1.7.1"
+    "socket.io": "^1.7.1",
+    "ws": "^1.1.1"
   },
   "devDependencies": {
+    "@types/express": "^4.0.34",
+    "@types/node": "^6.0.54",
+    "@types/socket.io": "^1.4.27",
+    "@types/ws": "0.0.37",
     "gulp": "^3.9.1",
     "gulp-tsc": "^1.2.5",
     "node-sass": "^3.13.0",
-    "typescript": "^2.0.10"
+    "typescript": "^2.1.4"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
